% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit2docs.R
\name{knit2docs}
\alias{knit2docs}
\title{Knit .Rmd File to Google Docs}
\usage{
knit2docs(rmd_file, doc_name = ".Rmd to Docs")
}
\arguments{
\item{rmd_file}{The file path of the .Rmd file that you want to knit.}

\item{doc_name}{The title of the Doc}
}
\description{
Knit a .Rmd file to a Google Doc (stored on Google Drive)
}
\examples{
# NOTRUN
knit2docs("Test.Rmd",doc_name=".Rmd to Docs")

}
