% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit2docs.R
\name{knit2docs}
\alias{knit2docs}
\title{Knit .Rmd File to Google Docs}
\usage{
knit2docs(rmd_file, doc_name = ".Rmd to Docs", path = NULL, overwrite = TRUE)
}
\arguments{
\item{rmd_file}{The file path of the .Rmd file that you want to knit.}

\item{doc_name}{The title of the Doc}

\item{path}{specify the path where the Google Doc will live. See the `path` argument in googledrive::drive_upload()}

\item{overwrite}{Overwrite file if it exists. See the `overwrite` argument in googledrive::drive_upload()}
}
\description{
Knit a .Rmd file to a Google Doc (stored on Google Drive)
}
\examples{
# NOTRUN
knit2docs("Test.Rmd",doc_name=".Rmd to Docs")

}
